+++ app/routes/dashboard/unused-advisor.js
@@ -2,7 +2,23 @@ import Ember from 'ember';
 import SidePanelRouteMixin from 'ember-cli-paint/mixins/side-panel-route';
 
 export default Ember.Route.extend(SidePanelRouteMixin, {
+  titleToken: function(model) {
+    let advisorName = model.get('advisor.name');
+
+    return `Unused Advisor: ${advisorName}`;
+  },
+
   model: function(params) {
-    return this.store.find('unusedAdvisor', params.unused_advisor_id);
+    return this.store.find('unusedAdvisor', params.unused_advisor_id).then((unusedAdvisor) => {
+      return Ember.RSVP.hash({
+        unusedAdvisor: unusedAdvisor,
+        projectHistory: this.store.find('projectHistory', { advisor_id: unusedAdvisor.get('advisor.id') })
+      });
+    });
+  },
+
+  setupController: function(controller, models) {
+    controller.set('model', models.unusedAdvisor);
+    controller.set('projectHistory', models.projectHistory);
   }
 });