+++ tests/test-helper.js
@@ -9,8 +9,6 @@ import {
 
 setResolver(resolver);
 
-QUnit.urlParams.nocontainer = true;
-
 export default {
   beforeEach: function() {
     this.app = startApp();