+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # Phoenix Changelog
 
+### Release 0.0.4 (June 15, 2015)
+
+- [#341](https://github.com/alphasights/phoenix/pull/341) Add team member filter to projects page
+- [#338](https://github.com/alphasights/phoenix/pull/338) Improve remove advisor confirmation flow
+- [#355 BUGFIX](https://github.com/alphasights/phoenix/pull/335) Don't trigger double confirmation for removing advisor from widget
+- [#306](https://github.com/alphasights/phoenix/pull/306) Add call completion form to interaction side panel
+- [#314](https://github.com/alphasights/phoenix/pull/314) Interactions now use the default project interaction type if present
+
 ### Release 0.0.3 (June 8, 2015)
 
 - [#315 BUGFIX](https://github.com/alphasights/phoenix/pull/315) Fix form submission errors not showing up correctly
@@ -9,6 +17,7 @@
 - [#301 BUGFIX](https://github.com/alphasights/phoenix/pull/301) Fix widget header and empty view copy bugs
 - [#300 BUGFIX](https://github.com/alphasights/phoenix/pull/300) Fix Team Members filter width
 - [#299](https://github.com/alphasights/phoenix/pull/299) Add basic version of Unused Advisors sidepanel
+- [#296](https://github.com/alphasights/phoenix/pull/296) Cancel Interaction from side panel
 
 ### Release 0.0.2 (June 1, 2015)